+++ src/pages/AddRecipesPage/AddRecipesPage.jsx
@@ -1,0 +1,18 @@
+import MainTitle from "../../components/MainTitle/MainTitle";
+import AddRecipeForm from "../../components/AddRecipeForm/AddRecipeForm";
+import FollowUs from "../../components/FollowUs/FollowUs";
+import PopularRecipe from "../../components/PopularRecipe/PopularRecipe";
+
+const AddRecipesPage = () => {
+  return (
+    <>
+      <div>AddRecipesPage</div>
+      <MainTitle></MainTitle>
+      <AddRecipeForm></AddRecipeForm>
+      <FollowUs></FollowUs>
+      <PopularRecipe></PopularRecipe>
+    </>
+  );
+};
+
+export default AddRecipesPage;
