+++ src/pages/CategoriesPage/CategoriesPage.jsx
