+++ src/pages/MyRecipesPage/MyRecipesPage.jsx
@@ -1,6 +1,6 @@
-import MainPageTitle from '../components/MainPageTitle/MainPageTitle'
-import MyRecipesList from '../components/MyRecipesList/MyRecipesList'
-import Paginator from '../components/Paginator/Paginator'
+import MainPageTitle from "../../components/MainPageTitle/MainPageTitle";
+import MyRecipesList from "../../components/MyRecipesList/MyRecipesList";
+import Paginator from "../../components/Paginator/Paginator";
 
 const MyRecipesPage = () => {
   return (
