+++ src/pages/SharedLayout/SharedLayout.jsx
@@ -1,9 +1,9 @@
 import { Outlet } from "react-router-dom";
 import { Suspense } from "react";
 
-import Header from "../components/Header/Header";
-import Footer from "../components/Footer/Footer";
-import Loader from "../components/Loader/Loader";
+import Header from "../../components/Header/Header";
+import Footer from "../../components/Footer/Footer";
+import Loader from "../../components/Loader/Loader";
 
 const SharedLayout = () => {
   return (
